連続して投稿や更新を行う際に発生するページの停止エラーを修正
연속 게시물 등록 및 업데이트시 페이지 멈추는 오류 수정
