+++ src/main/java/com/halo/admin/boardmanagement/notice/AdminNOTICEC.java
@@ -12,7 +12,6 @@ public class AdminNOTICEC extends HttpServlet {
 
 	protected void doGet(HttpServletRequest request, HttpServletResponse response)
 			throws ServletException, IOException {
-		System.out.println("진입확인 get");
 		NOTICEDAO.getAllNOTICE(request, response);
 		NOTICEDAO.NOTICEpagingAdmin(1, request);
 		
@@ -22,7 +21,6 @@ public class AdminNOTICEC extends HttpServlet {
 
 	protected void doPost(HttpServletRequest request, HttpServletResponse response)
 			throws ServletException, IOException {
-		System.out.println("진입확인 post");
 		NOTICEDAO.getAllNOTICE(request, response);
 
 		request.setAttribute("menu", "/admin/boardmanagement/notice/noticeContent.jsp");
