+++ src/main/java/com/halo/admin/boardmanagement/notice/Ck_DAO.java
@@ -4,9 +4,11 @@ import java.io.IOException;
 import java.io.UnsupportedEncodingException;
 import java.sql.Connection;
 import java.sql.PreparedStatement;
+import java.sql.ResultSet;
 
 import javax.servlet.http.HttpServletRequest;
 
+import com.halo.main.DBManagerhalo;
 import com.halo.test.DBManagerhalo_JW;
 
 public class Ck_DAO {
@@ -14,6 +16,10 @@ public class Ck_DAO {
 	public static void regNotice(HttpServletRequest request) throws IOException {
 
 		request.setCharacterEncoding("utf-8");
+		
+		Connection con = null;
+	    PreparedStatement pstmt = null;
+	    ResultSet rs = null;
 		
 		try {
 			String writer;
@@ -46,11 +52,6 @@ public class Ck_DAO {
 					}
 				}
 
-				int titleLength = title.length();
-				int txtLength = txt.length();
-
-				Connection con = null;
-				PreparedStatement pstmt = null;
 				String sql = "INSERT INTO announced_tbl (an_seq, an_title, an_content, an_writer, an_category) "
 						+ "VALUES (announced_tbl_seq.nextval, ?, ?, ?, ?)";
 
@@ -63,13 +64,15 @@ public class Ck_DAO {
 				pstmt.setString(4, select);
 
 				if (pstmt.executeUpdate() == 1) {
-					System.out.println("등록성공 Ck_DAO NOTICE");
-					System.out.println("---------------");
+					System.out.println("등록성공");
 				}
 			}
-
 		} catch (Exception e) {
 			e.printStackTrace();
+		} finally {
+				DBManagerhalo_JW.close(con, pstmt, rs);
+		        System.out.println("신규 등록 찐 닫음");
+		    
 		}
 
 	}
@@ -78,16 +81,19 @@ public class Ck_DAO {
 	public static void deleteNotice(HttpServletRequest request) throws UnsupportedEncodingException {
 
 		request.setCharacterEncoding("utf-8");
+		
+		Connection con = null;
+	    PreparedStatement pstmt = null;
+	    ResultSet rs = null;
+		
 		try {
-
 			String title = request.getParameter("title");
 			String select = request.getParameter("select");
 			String txt = request.getParameter("txt");
 			String seq = request.getParameter("seq");
-			System.out.println( "seq : !!!!!!!!이게 딜리트c라고 되있긴한데 업데이트를 하고있음" + seq);
+			
 			if (title != "" && txt != "" && select != null ) {
 
-				System.out.println("txt 구간(1) :::" + txt);
 				String[] saveFnameValues = request.getParameterValues("saveFname");
 
 				if (saveFnameValues != null) {
@@ -103,28 +109,9 @@ public class Ck_DAO {
 						txt = txt.substring(0, imgPos) + toReplace + txt.substring(txt.indexOf(">", imgPos) + 1);
 
 						startPos = imgPos + toReplace.length();
-						System.out.println();
-						System.out.println("txt 구간(2) :::" + txt);
-						System.out.println("saveFName: " + saveFnameValues[i]);
 					}
 				} // if
 
-				System.out.println("title : " + title);
-				System.out.println("select : " + select);
-				System.out.println("seq : " + seq);
-//			System.out.println(saveFName);
-//			txt = txt.replace("img", "img src=\'" + saveFName + "\'");
-
-				System.out.println("//////////////////////");
-				int titleLength = title.length();
-//			int selectLength = select.length();
-				int txtLength = txt.length();
-				System.out.println("제목 글자 수 : " + titleLength);
-//			System.out.println("카테고리 글자 수 : "+ selectLength);
-				System.out.println("내용 글자 수 : " + txtLength);
-
-				Connection con = null;
-				PreparedStatement pstmt = null;
 				String sql = "UPDATE announced_tbl SET an_title = ?, an_content = ?, an_category = ? WHERE an_seq = ?";
 
 				con = DBManagerhalo_JW.connect();
@@ -136,15 +123,15 @@ public class Ck_DAO {
 				pstmt.setString(4, seq);
 
 				if (pstmt.executeUpdate() == 1) {
-					System.out.println("수정성공 Ck_DAO NOTICE");
-					System.out.println("---------------");
+					System.out.println("수정성공");
 				}
 			}
-
 		} catch (Exception e) {
 			e.printStackTrace();
-		}
-
+		}  finally {
+			DBManagerhalo_JW.close(con, pstmt, rs);
+	        System.out.println("업데이트 찐닫음");
+	    }
 	}
 
 } // end class
