+++ src/main/java/com/halo/admin/boardmanagement/notice/CkeditorC.java
@@ -17,6 +17,7 @@ public class CkeditorC extends HttpServlet {
 	protected void doPost(HttpServletRequest request, HttpServletResponse response)
 			throws ServletException, IOException {
 		// 이곳은 Notice에서 등록완료를 눌렀을 때 등록하기 위한 곳
+		System.out.println("추가확인 11");
 		Ck_DAO.regNotice(request);
 	}
