+++ src/main/webapp/admin/boardmanagement/notice/noticeContent.jsp
@@ -59,7 +59,7 @@
 					<input hidden="1" name="aq_seq" value="${an.an_seq }">
 					<div class="ontent-m-td-2-content-txt-no-in">${(loop.index + 1) + (curPageNo - 1) * 8}</div>
 					<div class="ontent-m-td-2-content-txt-kategorie-in">${an.an_category }</div>
-					<div class="ontent-m-td-2-content-txt-title-in"><a onclick="getNOTICEDataV('${an.an_seq}')">${an.an_title }</a></div>
+					<div class="ontent-m-td-2-content-txt-title-in"><a onclick="getNoticeViewData('${an.an_seq}')">${an.an_title }</a></div>
 					<div class="ontent-m-td-2-content-txt-writer-in">${an.an_writer }</div>
 					<div class="ontent-m-td-2-content-txt-date-in">${an.an_reg_date}</div>
 					<div class="ontent-m-td-2-content-txt-delete-in">
